Extract trajectory frame analysis into $lib/trajectory/analysis
