+++ src/lib/structure-id/TrajectoryStructureIdPane.svelte
@@ -2,7 +2,7 @@
   import { Cross, Lattice } from 'svelte-widgets/icons'
   import { StatusMessage } from '$lib/feedback'
   import { format_num } from '$lib/labels'
-  import { analysis_pane_setup } from '$lib/msd/collect'
+  import { analysis_pane_setup } from '$lib/trajectory/analysis'
   import { DraggablePane, type PaneProps, type PaneToggleProps } from '$lib/overlays'
   import type { TrajectoryType } from '$lib/trajectory'
   import { to_error } from '$lib/utils'
