+++ src/lib/vacf/TrajectoryVacfPane.svelte
@@ -3,7 +3,7 @@
   import { StatusMessage } from '$lib/feedback'
   import { WINDOW_TYPES, type WindowType } from '$lib/fft'
   import { format_bytes, format_num } from '$lib/labels'
-  import { analysis_pane_setup } from '$lib/msd/collect'
+  import { analysis_pane_setup } from '$lib/trajectory/analysis'
   import { DraggablePane, type PaneProps, type PaneToggleProps } from '$lib/overlays'
   import type { ParseProgress, TrajectoryType } from '$lib/trajectory'
   import { to_error } from '$lib/utils'
